[Decomp][Script] Cohesion split: extract script binding to ScriptMgrBinding.cpp
